Major refactor, all js converted to ts
Initally this work added the ability to enable and disable the automatic
a11y checking on DOM updates. Along the way, I realized I could simplify
the internal/external panel handling. That refactor was large enough that
I went ahead and converted the js to TypeScript to make the process easier.